Final Project - Cyberbullying Detection/backend/requirements.txt
gunicorn
requests
scikit-learn==0.24.1
flask==2.0.2
pandas
tensorflow-cpu
numpy